Unit Test amd Lint Error fixed
