+++ .babelrc
@@ -3,9 +3,7 @@
     [
       "@babel/preset-env",
       {
-        "targets": {
-          "node": true
-        }
+        "modules": false
       }
     ],
     "@babel/preset-react",
