+++ jest.config.ts
@@ -186,7 +186,8 @@ export default {
   // timers: "real",
 
   // A map from regular expressions to paths to transformers
-  transform: {
+  transform:  {
+    '^.+\\.(ts|tsx)$': 'ts-jest',
     '^.+\\.tsx?$': 'babel-jest',
   },
