+++ package.json
@@ -1,6 +1,7 @@
 {
   "name": "@modusoperandi/licit-paste-json",
   "version": "0.0.12",
+  "type": "module",
   "description": "MO's ProseMirror Plugin that allows you to paste in prosemirror JSON into the editor.",
   "main": "dist/index.js",
   "typings": "./index.d.ts",
@@ -26,8 +27,12 @@
     "@babel/plugin-proposal-export-default-from": "^7.18.10",
     "@babel/plugin-proposal-export-namespace-from": "^7.18.9",
     "@babel/plugin-proposal-function-sent": "^7.18.6",
+    "@babel/plugin-proposal-json-strings": "^7.18.6",
     "@babel/plugin-proposal-logical-assignment-operators": "^7.18.9",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.15.0",
+    "@babel/plugin-proposal-numeric-separator": "^7.15.0",
     "@babel/plugin-proposal-object-rest-spread": "^7.20.2",
+    "@babel/plugin-proposal-optional-chaining": "^7.15.0",
     "@babel/plugin-proposal-pipeline-operator": "^7.18.9",
     "@babel/plugin-proposal-throw-expressions": "^7.18.6",
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
@@ -35,19 +40,15 @@
     "@babel/plugin-transform-parameters": "^7.20.5",
     "@babel/plugin-transform-runtime": "^7.19.6",
     "@babel/plugin-transform-typescript": "^7.20.2",
-    "@babel/plugin-proposal-json-strings": "^7.18.6",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.15.0",
-    "@babel/plugin-proposal-numeric-separator": "^7.15.0",
-    "@babel/plugin-proposal-optional-chaining": "^7.15.0",
     "@babel/preset-env": "^7.20.2",
     "@babel/preset-react": "^7.18.6",
     "@babel/preset-typescript": "^7.18.6",
     "@cyclonedx/cyclonedx-npm": "^1.11.0",
     "@types/jest": "^29.2.4",
+    "@types/node": "^18.7.22",
     "@typescript-eslint/eslint-plugin": "^5.46.1",
     "@typescript-eslint/parser": "^5.46.1",
     "babel-loader": "^9.1.0",
-    "@types/node": "^18.7.22",
     "babel-plugin-transform-react-remove-prop-types": "0.4.24",
     "clean-webpack-plugin": "^4.0.0",
     "css-loader": "^6.7.3",
