+++ webpack.config.js
@@ -1,15 +1,18 @@
 /* eslint-disable */
 
-var webpack = require('webpack'),
-  TerserPlugin = require('terser-webpack-plugin'),
-  WriteFilePlugin = require('write-file-webpack-plugin'),
-  path = require('path');
-
-const {CleanWebpackPlugin} = require('clean-webpack-plugin');
+import webpack from 'webpack';
+import TerserPlugin from 'terser-webpack-plugin';
+import WriteFilePlugin from 'write-file-webpack-plugin';
+import path, {dirname} from 'path';
+import {CleanWebpackPlugin} from 'clean-webpack-plugin';
+import {fileURLToPath} from 'url';
 
 const NODE_ENV = process.env.NODE_ENV || 'production';
 
 var isDev = 'development' === NODE_ENV || 0;
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
 var options = {
   mode: NODE_ENV,
@@ -18,10 +21,18 @@ var options = {
   },
   output: {
     path: path.join(__dirname, 'bin'),
-    filename: '[name].bundle.js'
+    filename: '[name].bundle.js',
   },
   module: {
     rules: [
+      {
+        test: /\.(js|mjs|jsx)$/,
+        include: /node_modules/,
+        type: 'javascript/auto',
+        resolve: {
+          fullySpecified: false,
+        },
+      },
       {
         test: /\.(ts|tsx)$/,
         exclude: /node_modules/,
@@ -29,20 +40,17 @@ var options = {
       },
       {
         test: /\.css$/,
-        use: [
-          'style-loader',
-          'css-loader',
-        ],
+        use: ['style-loader', 'css-loader'],
       },
       {
-        test: /\.(jpe?g|png|gif|svg)$/i, 
-        loader: 'file-loader'
+        test: /\.(jpe?g|png|gif|svg)$/i,
+        loader: 'file-loader',
       },
     ],
   },
   resolve: {
     alias: {},
-    extensions: ['.ts','.js','.json']
+    extensions: ['.ts', '.js', '.json'],
   },
   plugins: [
     // clean the web folder
@@ -51,7 +59,7 @@ var options = {
     new webpack.DefinePlugin({
       'process.env.NODE_ENV': JSON.stringify(NODE_ENV),
     }),
-    new WriteFilePlugin()
+    new WriteFilePlugin(),
   ],
 };
 
@@ -77,4 +85,4 @@ options.plugins.push(function () {
   });
 });
 
-module.exports = options;
+export default options;
